// Web/RMA-Web-Shell/projects/fincare/src/app/payment-manager/models/payment.model.ts
import { BaseClass } from 'projects/shared-models-lib/src/lib/common/base-class';

export class Payment extends BaseClass {
    paymentId: number;
    paymentStatus: number;
    paymentType: number;
    payee: string;
    bank: string;
    bankBranch: string;
    accountNo: string;
    amount: number;
    retainedCommission: string;
    product: string;
    company: number;
    branch: number;
    senderAccountNo: string;
    brokerCode: number;
    brokerName: string;
    fsbAccredited: boolean;
    errorCode: string;
    maxSubmissionCount: number;
    submissionCount: number;
    bankAccountType: number;
    idNumber: string;
    emailAddress: string;
    paymentTypeDesc: string;
    paymentStatusDesc: string;
    errorDescription: string;
    claimTypeId: number;
    accountTypeDesc: string;
    claimTypeDesc: string;
    submissionDate: Date;
    clientNotificationDate: Date;
    paymentConfirmationDate: Date;
    rejectionDate: Date;
    reconciliationDate: Date;
    recalledDate: Date;
    facsTransactionResults: [any];
    dialogQuestion: string;
    dialogView: string;
    payment: any;
    canResubmit: boolean;
    clientType: number;
    reference: string;
    claimReference: string;
    batchReference: string;
    policyReference: string;
    paymentSubmissonBatchid: number;
    noOfTransactions: number;
    policyId: number;
    period: string;
    isBetweenTenAndTwo: boolean;
    bankStatementReference: string;
    strikeDate: Date;
    memberName: string;
    memberNumber: string;
    benefitType: number;
    scheme: string;
    currency:string;
    isForex: boolean;
}